modifiche per plottare con array float

+++ lib/estrazione_csv/include/estrattore_csv.h
@@ -21,7 +21,7 @@
         float altezza_mozzo;
 		float *wind_speed;
         float *power_coefficients;
-        int *power_curves;
+        float *power_curves;
         struct turbina *prev;
     };
 
@@ -90,6 +90,6 @@
 	
 	void reading_file_power_curves(struct csv *file, struct turbina *const puntatore, char *percorso_file_power_curves, int *errore);
 	
-	void inserimento_power_curves(int *array_dati, char **fields);
+	void inserimento_power_curves(float *array_dati, char **fields);
     
 #endif
